final commit

+++ index.php
@@ -2,8 +2,8 @@
 
 require_once("calc/calculator.php");
 
-$numb1 = readline("Qual o primeiro número:");
-$numb2 = readline("Qual o segundo número:");
+$numb1 = (float)readline("Qual o primeiro número:");
+$numb2 = (float)readline("Qual o segundo número:");
 $op = readline("Qual a Operação a realizar (ex: '+','-','/','*'):");
 $cal = $numb1.$op.$numb2;
 try {
